Marian now builds the Visual Studio CPU-only versions, Release and Debug, without warnings.

Only two warnings had to be disabled globally (unreachable code in some STL header, deprecated POSIX functions).

The majority were implicit typecasts e.g. size_t to int or float to int.

It's a lot of files, but most have only one change. Do the PR in CodeFlow, it will be much quicker than the web interface.

Related work items: #93467

+++ src/data/shortlist.h
@@ -79,7 +79,7 @@ public:
       for(auto i : srcBatch->data())
         idxSet.insert(i);
 
-    std::uniform_int_distribution<> dis(firstNum_, maxVocab_);
+    std::uniform_int_distribution<> dis((int)firstNum_, (int)maxVocab_);
     while(idxSet.size() < total_ && idxSet.size() < maxVocab_)
       idxSet.insert(dis(gen_));
